update exception handling. Add docstrings

+++ utils/gcp_bucket.py
@@ -12,6 +12,9 @@ logging.basicConfig(level=logging.DEBUG)
 
 
 class CloudStorageBucketClient:
+    """
+    Client class used for connectiong Google Cloud Storage
+    """
     def __init__(self):
         try:
             self.__storage_client = storage.Client()
@@ -23,6 +26,13 @@ class CloudStorageBucketClient:
         return self.__storage_client
 
     def create_bucket(self, bucket_name, location=GCP_LOCATION_EU):
+        """
+        Creates the bucket in the Google cloud storage
+        :param bucket_name: name of the bucket to be created
+        :param location: physical location of servers in which
+        this bucket must be
+        :return: recently created Bucket object
+        """
         try:
             self.storage_client.create_bucket(bucket_name, location=location)
             logging.info('Bucket created successfully')
@@ -38,6 +48,12 @@ class CloudStorageBucketClient:
             return 'Creating new bucket failed'
 
     def retrieve_bucket(self, bucket_name):
+        """
+        Returns the bucket with the given name
+
+        :param bucket_name:  name of the bucket to be retrieved
+        :return: Bucket object with the given name
+        """
         try:
             bucket = self.storage_client.get_bucket(bucket_name)
             logging.info('Bucket retrieved successfully')
@@ -47,8 +63,19 @@ class CloudStorageBucketClient:
             logging.debug(f'GoogleAPIError occurred: {google_api_error}')
 
             return f'Retrieving bucket with name {bucket_name} failed'
+        except ValueError as value_error:
+            logging.debug(f'Invalid name: {value_error}')
+
+            return f'Retrieving bucket with name {bucket_name} failed'
 
     def retrieve_buckets(self, max_results=None, prefix=None):
+        """
+        Lists the buckets available in that particular project
+
+        :param max_results: the number of buckets to be returned
+        :param prefix: buckets with this prefix
+        :return: all the buckets available in that particular project
+        """
         try:
             buckets = self.storage_client.list_buckets(max_results=max_results,
                                                        prefix=prefix)
@@ -61,8 +88,20 @@ class CloudStorageBucketClient:
             return 'Retrieving buckets failed'
 
     def delete_bucket(self, bucket_name, force_deletion=False):
+        """
+        Deletes the bucket with the given name
+
+        :param bucket_name: name of the bucket to be deleted
+        :param force_deletion: flag whether to force to delete non-empty
+        bucket or not
+        :return: #TODO ???
+        """
         try:
             bucket = self.retrieve_bucket(bucket_name)
+
+            if type(bucket) == 'str':
+                raise ValueError
+
             bucket.delete(force=force_deletion)
 
             return 'Deleted', 204
@@ -70,3 +109,6 @@ class CloudStorageBucketClient:
             logging.debug(f'GoogleAPIError occurred: {google_api_error}')
 
             return f'Deleting bucket with name {bucket_name} failed'
+        except AttributeError:
+            logging.debug(f'Invalid name or'
+                          f' Bucket with that name does not exist')
